feat: lab-4

+++ Kotiki-lab-3/src/main/java/ru/itmo/main/service/CatService.java
@@ -1,16 +1,16 @@
 package ru.itmo.main.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
-import ru.itmo.main.dao.model.CatsEntity;
-import ru.itmo.main.dao.model.Color;
-import ru.itmo.main.dao.model.FriendsOfCatsEntity;
-import ru.itmo.main.dao.model.OwnersEntity;
+import ru.itmo.main.dao.model.*;
 import ru.itmo.main.dao.repository.CatRepository;
 import ru.itmo.main.dao.repository.FriendRepository;
+import ru.itmo.main.dao.repository.UserRepository;
 
 import java.sql.Date;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class CatService {
@@ -21,11 +21,24 @@ public class CatService {
     @Autowired
     private FriendRepository friendRepository;
 
+    @Autowired
+    private UserRepository userRepository;
+
+    private boolean isAdmin(){
+        return SecurityContextHolder.getContext().getAuthentication().getAuthorities().stream().anyMatch(s -> s.equals(Role.ADMIN.toString()));
+    }
+
+    private UserEntity getUser(){
+        return userRepository.findByUserName(SecurityContextHolder.getContext().getAuthentication().getName());
+    }
+
+
     public CatsEntity saveCat(String name,
                               String breed,
                               Color color,
                               OwnersEntity owner,
-                              Date birthday){
+                              Date birthday) throws Exception {
+        if(!owner.equals(getUser().getUserOwner())) throw new Exception("It's not your owner");
         CatsEntity catsEntity = new CatsEntity();
         catsEntity.setName(name);
         catsEntity.setBreed(breed);
@@ -37,19 +50,28 @@ public class CatService {
     }
 
     public List<CatsEntity> getAllCats(){
-        return catRepository.findAll();
+        if(isAdmin()) {
+            return catRepository.findAll();
+        }
+        return getUser().getUserOwner().getCats().stream().collect(Collectors.toList());
     }
 
     public List<CatsEntity> getAllCatsByColor(Color color){
-        return catRepository.findAllByColor(color);
+        if(isAdmin())
+            return catRepository.findAllByColor(color);
+        return catRepository.findAllByColorAndOwner(color, getUser().getUserOwner());
     }
 
-    public CatsEntity getCatById(int id){
-        return catRepository.findById(id);
+    public CatsEntity getCatById(int id) throws Exception {
+        CatsEntity catsEntity = catRepository.findById(id);
+        if(isAdmin() || catsEntity.getOwner().equals(getUser().getUserOwner()))
+            return catsEntity;
+        throw new Exception("It's not your cat");
     }
 
     public void deleteCat(CatsEntity catsEntity){
-        catRepository.delete(catsEntity);
+        if(isAdmin() || catsEntity.getOwner().equals(getUser().getUserOwner()))
+         catRepository.delete(catsEntity);
     }
 
     public void deleteCatById(int id){
@@ -60,11 +82,11 @@ public class CatService {
         catRepository.saveAndFlush(catsEntity);
     }
 
-    public List<CatsEntity> getFriends(int id){
+    public List<CatsEntity> getFriends(int id) throws Exception {
         return getCatById(id).getFriends();
     }
 
-    public void makeFriends(int cat1, int cat2){
+    public void makeFriends(int cat1, int cat2) throws Exception {
         if(getFriends(cat1).stream().anyMatch(cat -> cat.getId() == cat2)){
             return;
         }
